Refactoring to remove warnings due to scalastyle.
This commits add modifications to remove the warnings due to scalastyle
plugin. Most of the style related stuff is handled by scalariform plugin
for intellij. At this point the preview example is working and is not
thoroughly tested for all the changes. However the style changes should
not break anything.